requirements.txt
requests
pytest
selenium
urllib3
allure-pytest
allure-python-commons
parameterized